remove metric subtypes in favor of one general one. move metric and parsing code to its own module. move client to its own module. formalize creation of metrics from packet
